requirements.txt
# =============================================================================
# Gastro-Savdo-Insights - Python Dependencies
# =============================================================================
# 
# O'rnatish: pip install -r requirements.txt
# Virtual environment yaratish: python -m venv venv
# Aktivatsiya: source venv/bin/activate (macOS/Linux)
#              .\venv\Scripts\activate (Windows)
# =============================================================================

# ==================== Core Framework ====================
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2

# ==================== Database ====================
mysql-connector-python==9.0.0
PyMySQL==1.1.1
sqlalchemy==2.0.23

# ==================== Security ====================
bcrypt==4.2.0
PyJWT==2.9.0
python-multipart==0.0.9

# ==================== HTTP Client (Telegram API uchun) ====================
httpx==0.27.2

# ==================== Environment ====================
python-dotenv==1.0.1

# ==================== Development ====================
# pytest==8.3.2
# pytest-asyncio==0.24.0
# httpx==0.27.2  # Test client uchun ham

# ==================== Logging & Monitoring ====================
# loguru==0.7.2

# ==================== Optional: Redis (Production 2FA uchun) ====================
# redis==5.0.8
# aioredis==2.0.1
